Replace consume and produce* functions with transduce
Each call to transduce expresses the transformation of a single chunk
of source content. transduce takes a count of characters to consume,
optional content, and an optional traversal extent. If no content is
provided, we pass through the consumed source content. The optional
traversal extent replaces the `produceNewline` and `produceCharacter`
methods, allowing the effects of both to be expressed in spatial terms.
The advantage here is a simpler interface and less state required to
align source and target traversals.