Remove TransformIterator and just use TransformBuffer directly
In reality, the TransformBuffer isn’t *quite* an iterator in the same
sense that the other iterators are, because it’s not really traversing
through an existing structure (it has a reset method instead of seek).

Seems like trying to treat it as an iterator leads to more confusion
than clarity, so lets put it back how it was. Still like it in a
different file though.

/cc @maxbrunsfeld