WIP: Include leaf nodes in path array
@maxbrunsfeld: Take a look at the new approach. I started storing the
offset at each entry of the path (formerly stack) because I think it’s
needed in order to adjust the extent of the parent nodes.

This passes the patch specs but there’s lots of issues elsewhere. I 
think this is the right direction but the git history is all there if
you disagree.